pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "budget"
version = "0.1.0"
description = "Spectral energy budget of a non-hydrostatic atmosphere"
readme = "README.md"
authors = [{ name = "Bharath Krishnan", email = "krishnan@iap-kborn.de" },
    { name = "Yanmichel A. Morfa", email = "morfa@iap-kborn.de" }]
requires-python = ">=3.10"
dependencies = [
    "xarray>=2024.1.0",
    "dask[array]>=2024.5.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "zarr>=2.17",
    "h5netcdf>=1.3",
    "netcdf4>=1.6"
]

[project.scripts]
budget = "budget.client:main"